pyproject.toml

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoops"
version = "0.1.0"
description = "AutoOps - AI DevOps Agent (Groq backend)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.22.0",
    "typer[all]>=0.7.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "groq>=0.2.0"
]

[project.scripts]
autoops = "autoops.cli:app"
